@@ -26,7 +26,7 @@
         "tsserver": "./bin/tsserver"
     },
     "engines": {
-        "node": ">=4.2.0"
+        "node": ">=14.17"
     },
     "files": [
         "bin",
@@ -39,75 +39,63 @@
         "!**/.gitattributes"
     ],
     "devDependencies": {
-        "@octokit/rest": "latest",
-        "@types/chai": "latest",
-        "@types/fancy-log": "^2.0.0",
-        "@types/fs-extra": "^9.0.13",
-        "@types/glob": "latest",
-        "@types/gulp": "^4.0.9",
-        "@types/gulp-concat": "latest",
-        "@types/gulp-newer": "latest",
-        "@types/gulp-rename": "latest",
-        "@types/gulp-sourcemaps": "latest",
-        "@types/merge2": "latest",
-        "@types/microsoft__typescript-etw": "latest",
-        "@types/minimist": "latest",
-        "@types/mkdirp": "latest",
-        "@types/mocha": "latest",
-        "@types/ms": "latest",
+        "@esfx/canceltoken": "^1.0.0",
+        "@octokit/rest": "^19.0.13",
+        "@types/chai": "^4.3.4",
+        "@types/glob": "^8.1.0",
+        "@types/microsoft__typescript-etw": "^0.1.1",
+        "@types/minimist": "^1.2.2",
+        "@types/mocha": "^10.0.1",
+        "@types/ms": "^0.7.31",
         "@types/node": "latest",
-        "@types/source-map-support": "latest",
+        "@types/source-map-support": "^0.5.6",
         "@types/which": "^2.0.1",
-        "@types/xml2js": "^0.4.11",
-        "@typescript-eslint/eslint-plugin": "^5.33.1",
-        "@typescript-eslint/parser": "^5.33.1",
-        "@typescript-eslint/utils": "^5.33.1",
-        "azure-devops-node-api": "^11.2.0",
-        "chai": "latest",
+        "@typescript-eslint/eslint-plugin": "^6.0.0",
+        "@typescript-eslint/parser": "^6.0.0",
+        "@typescript-eslint/utils": "^6.0.0",
+        "azure-devops-node-api": "^12.0.0",
+        "c8": "^7.14.0",
+        "chai": "^4.3.7",
         "chalk": "^4.1.2",
-        "del": "^6.1.1",
+        "chokidar": "^3.5.3",
         "diff": "^5.1.0",
+        "dprint": "^0.42.3",
+        "esbuild": "^0.19.0",
         "eslint": "^8.22.0",
         "eslint-formatter-autolinkable-stylish": "^1.2.0",
-        "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-jsdoc": "^39.3.6",
         "eslint-plugin-local": "^1.0.0",
         "eslint-plugin-no-null": "^1.0.2",
-        "fancy-log": "latest",
-        "fs-extra": "^9.1.0",
-        "glob": "latest",
-        "gulp": "^4.0.2",
-        "gulp-concat": "latest",
-        "gulp-insert": "latest",
-        "gulp-newer": "latest",
-        "gulp-rename": "latest",
-        "gulp-sourcemaps": "latest",
-        "merge2": "latest",
-        "minimist": "latest",
-        "mkdirp": "latest",
-        "mocha": "latest",
-        "mocha-fivemat-progress-reporter": "latest",
+        "eslint-plugin-simple-import-sort": "^10.0.0",
+        "fast-xml-parser": "^4.0.11",
+        "glob": "^8.1.0",
+        "hereby": "^1.6.4",
+        "jsonc-parser": "^3.2.0",
+        "minimist": "^1.2.8",
+        "mocha": "^10.2.0",
+        "mocha-fivemat-progress-reporter": "^0.1.0",
         "ms": "^2.1.3",
         "node-fetch": "^3.2.10",
-        "source-map-support": "latest",
-        "typescript": "^4.8.4",
-        "vinyl": "latest",
-        "which": "^2.0.2",
-        "xml2js": "^0.4.23"
+        "playwright": "^1.38.0",
+        "source-map-support": "^0.5.21",
+        "tslib": "^2.5.0",
+        "typescript": "^5.0.2",
+        "which": "^2.0.2"
     },
     "overrides": {
-        "es5-ext": "0.10.53"
+        "typescript@*": "$typescript"
     },
     "scripts": {
-        "test": "gulp runtests-parallel --light=false",
-        "test:eslint-rules": "gulp run-eslint-rules-tests",
+        "test": "hereby runtests-parallel --light=false",
+        "test:eslint-rules": "hereby run-eslint-rules-tests",
         "build": "npm run build:compiler && npm run build:tests",
-        "build:compiler": "gulp local",
-        "build:tests": "gulp tests",
+        "build:compiler": "hereby local",
+        "build:tests": "hereby tests",
+        "build:tests:notypecheck": "hereby tests --no-typecheck",
         "start": "node lib/tsc",
-        "clean": "gulp clean",
-        "gulp": "gulp",
-        "lint": "gulp lint",
+        "clean": "hereby clean",
+        "gulp": "hereby",
+        "lint": "hereby lint",
+        "format": "dprint fmt",
         "setup-hooks": "node scripts/link-hooks.mjs"
     },
     "browser": {
@@ -118,11 +106,12 @@
         "buffer": false,
         "@microsoft/typescript-etw": false,
         "source-map-support": false,
-        "inspector": false
+        "inspector": false,
+        "perf_hooks": false
     },
-    "packageManager": "npm@8.15.0",
+    "packageManager": "npm@8.19.4",
     "volta": {
-        "node": "14.20.0",
-        "npm": "8.15.0"
+        "node": "20.1.0",
+        "npm": "8.19.4"
     }
 }
