@@ -1,26 +1,36 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+import React from 'react';
+import ReactDOM from 'react-dom';
 import {
-  BrowserRouter as Router,
-  Route,
-  Switch,
-  Redirect,
-} from 'react-router-dom'
+    BrowserRouter as Router,
+    Routes,
+    Route,
+    Navigate,
+} from 'react-router-dom';
 
-import './style.css'
-import Home from './views/home'
-import NotFound from './views/not-found'
+import './styles/style.css';
+import Home from './views/home';
+import NotFound from './views/not-found';
+import Dashboard from './views/Dashboard';
+import Header from './components/Header';
+import LoginPageComponent from './components/LoginPageComponent';
 
-const App = () => {
-  return (
-    <Router>
-      <Switch>
-        <Route component={Home} exact path="/" />
-        <Route component={NotFound} path="**" />
-        <Redirect to="**" />
-      </Switch>
-    </Router>
-  )
-}
+const App: React.FC = () => {
+    return (
+        <Router>
+            <div>
+                <Header />
+                <div style={{ paddingTop: '96px' }}>
+                    <Routes>
+                        <Route element={<Home />} path="/" />
+                        <Route element={<Dashboard />} path="/dashboard" />
+                        <Route element={<LoginPageComponent />} path="/login" />
+                        {/* Route for NotFound - matches any unknown route */}
+                        <Route element={<NotFound />} path ="*" />
+                    </Routes>
+                </div>
+            </div>
+        </Router>
+    );
+};
 
-ReactDOM.render(<App />, document.getElementById('app'))
+ReactDOM.render(<App />, document.getElementById('app'));
