@@ -2,7 +2,7 @@ import React from 'react'
 
 import PropTypes from 'prop-types'
 
-import './feature-card.css'
+import '../styles/feature-card.css'
 
 const FeatureCard = (props) => {
   return (
